+++ src/TwigUtilities/Functions/CopyrightYear.php
@@ -8,11 +8,12 @@ class CopyrightYear extends AbstractFunction
 {
   /**
    * @param int|string $fromYear starting year
+   * @param int|string $endYear  ending year. Defaults to current year
    */
-  public function __invoke($fromYear): string
+  public function __invoke($fromYear, $endYear = "now"): string
   {
     $startY = \DateTime::createFromFormat("Y", "{$fromYear}")->format("Y");
-    $endY = date("Y");
+    $endY = \DateTime::createFromFormat("Y", "{$endYear}")->format("Y");
 
     return ($startY >= $endY) ? $startY : "{$startY}-{$endY}";
   }