+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.9.0] - 2020-04-14
+### Added
+- StaticTwigPage Slim action.
+- Some tests.
+
+### Changed
+- `CopyrightYear`: Now accepts an optional end date.
+- `FormatDateRange`: Add period at the end of dates without a year.
+
 ## [0.8.0] - 2020-02-11
 This release introduces multiple major breaking changes. See updated
 [README](README.md) for new usage instructions.